Split codeflower into two files
